+++ interface/src/reloading.rs
@@ -7,10 +7,10 @@ use std::sync::atomic::{AtomicPtr, Ordering::*};
 #[derive(Clone, Copy, Debug, PartialEq, Eq)]
 pub struct Functions {
     pub render: unsafe fn(*mut c_void,  &mut Graphics),
-    pub update: unsafe fn(*mut c_void,  f64),
+    pub update: unsafe fn(*mut c_void,  f32),
     pub key_press: unsafe fn(*mut c_void,  Key),
     pub key_release: unsafe fn(*mut c_void,  Key),
-    pub mouse_move: unsafe fn(*mut c_void,  [f64; 2]),
+    pub mouse_move: unsafe fn(*mut c_void,  [f32; 2]),
     pub mouse_press: unsafe fn(*mut c_void,  MouseButton),
     pub size: usize
 }
@@ -41,7 +41,7 @@ impl Game for ReloadableGame {
     fn render(&mut self,  gfx: &mut Graphics) {
         unsafe{ (self.get().render)(self.game, gfx) };
     }
-    fn update(&mut self,  dt: f64) {
+    fn update(&mut self,  dt: f32) {
         unsafe{ (self.get().update)(self.game, dt) };
     }
     fn key_press(&mut self,  key: Key) {
@@ -53,7 +53,7 @@ impl Game for ReloadableGame {
     fn mouse_press(&mut self,  button: MouseButton) {
         unsafe{ (self.get().mouse_press)(self.game, button) };
     }
-    fn mouse_move(&mut self,  pos: [f64; 2]) {
+    fn mouse_move(&mut self,  pos: [f32; 2]) {
         unsafe{ (self.get().mouse_move)(self.game, pos) };
     }
 }
@@ -74,7 +74,7 @@ macro_rules! expose_game_reloadably{($dir:literal/$mod:tt::$game:tt = $target:li
     unsafe fn game_render_dyn(gamestate: *mut c_void,  g: &mut Graphics) {
         (&mut*(gamestate as *mut $game)).render(g)
     }
-    unsafe fn game_update_dyn(gamestate: *mut c_void,  deltatime: f64) {
+    unsafe fn game_update_dyn(gamestate: *mut c_void,  deltatime: f32) {
         (&mut*(gamestate as *mut $game)).update(deltatime)
     }
     unsafe fn game_key_press_dyn(gamestate: *mut c_void,  key: Key) {
@@ -83,7 +83,7 @@ macro_rules! expose_game_reloadably{($dir:literal/$mod:tt::$game:tt = $target:li
     unsafe fn game_key_release_dyn(gamestate: *mut c_void,  key: Key) {
         (&mut*(gamestate as *mut $game)).key_release(key)
     }
-    unsafe fn game_mouse_move_dyn(gamestate: *mut c_void,  pos: [f64;2]) {
+    unsafe fn game_mouse_move_dyn(gamestate: *mut c_void,  pos: [f32;2]) {
         (&mut*(gamestate as *mut $game)).mouse_move(pos)
     }
     unsafe fn game_mouse_press_dyn(gamestate: *mut c_void,  button: MouseButton) {
